src/main/java/com/uff/qat/intro_junit/produto/Produto.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.uff.qat.intro_junit.produto;

import java.util.Objects;

/**
 *
 * @author luizlaljr
 */
public class Produto {
    
    private String nome;
    private double preco;

    public Produto(String nome, double preco) {
        super();
        this.nome = nome;
        this.preco = preco;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public double getPreco() {
        return preco;
    }

    public void setPreco(double preco) {
        this.preco = preco;
    }

    @Override
    public boolean equals(Object p) {
        return ((Produto) p).getNome().equals(this.getNome());
    }
     
}
